Adding showDialog()
this code is very messy as i was learning how to implement logging with its package. now i am implementing exploring what would be best to use in showing a mini window that is triggered with a button click.

for now i will use the method called showDialog()

more to be followed...